Updates to RenderSystem
RenderSystem now batches everything and render using only one draw call.

This is done by the help of using Bindless Textures, a method of sampling the data based on a handle id instead of binding it based on the OpenGL context.

RenderSystem now no longer render any basic shapes. Everything is a sprite. If not file path to an image is found, a default white image will be loaded in instead.