Code for audio
Code have yet to be tested, however this is what it is supposed to do.

The main idea for the audio manager is that every frame, it will loop through a list of channels that are currently bound to a sound. A channel will be unavailable for used as long as it is still playing. When the sound stops playing, it will be added to the appropriate queue where it will store all the available channels for used. If no channels are available, no audio will be played.

Created function interface to interact with the audio.

Will need to continue my progress by adding codes to adjust the volume via the ChannelGroup functionality provided by fmod